+++ clang/lib/Lex/PPMacroExpansion.cpp
@@ -1741,7 +1741,7 @@ void Preprocessor::ExpandBuiltinMacro(Token &Tok) {
          getLangOpts().C2y ? diag::warn_counter : diag::ext_counter);
     // __COUNTER__ expands to a simple numeric value that must be less than
     // 2147483647.
-    constexpr unsigned long MaxPosValue = std::numeric_limits<int32_t>::max();
+    constexpr uint32_t MaxPosValue = std::numeric_limits<int32_t>::max();
     if (CounterValue > MaxPosValue) {
       Diag(Tok.getLocation(), diag::err_counter_overflow);
       // Retain the maximal value so we don't issue conversion-related
